+++ src/mc_tasks/PositionBasedVisServoTask.cpp
@@ -84,16 +84,17 @@ static auto registered = mc_tasks::MetaTaskLoader::register_load_function(
     "pbvs",
     [](mc_solver::QPSolver & solver, const mc_rtc::Configuration & config) {
       std::shared_ptr<mc_tasks::PositionBasedVisServoTask> t;
+      auto robotIndex = robotIndexFromConfig(config, solver.robots(), "pbvs");
       if(config.has("surface"))
       {
         t = std::make_shared<mc_tasks::PositionBasedVisServoTask>(config("surface"), sva::PTransformd::Identity(),
-                                                                  solver.robots(), config("robotIndex"));
+                                                                  solver.robots(), robotIndex);
       }
       else if(config.has("body"))
       {
         t = std::make_shared<mc_tasks::PositionBasedVisServoTask>(config("body"), sva::PTransformd::Identity(),
                                                                   config("X_b_s", sva::PTransformd::Identity()),
-                                                                  solver.robots(), config("robotIndex"));
+                                                                  solver.robots(), robotIndex);
       }
       t->load(solver, config);
       return t;